pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"

requires = ["setuptools>=61", "wheel"]

[project]
name = "comfyui-nunchaku"
version = "1.0.1"
description = "Nunchaku ComfyUI Node. Nunchaku is a high-performance inference engine optimized for low-bit neural networks. See more details: https://github.com/nunchaku-tech/nunchaku"
license = { file = "LICENCE.txt" }
requires-python = ">=3.10,<3.14"
#  Used by Comfy Registry https://comfyregistry.org
classifiers = [
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "accelerate>=1.10",
  "diffusers>=0.35",
  "huggingface-hub>=0.34",
  "peft>=0.17",
  "protobuf",
  "sentencepiece",
  "tomli",
]

optional-dependencies.compile = ["nunchaku @ git+https://github.com/nunchaku-tech/nunchaku.git"]
optional-dependencies.pulid = [
  "facexlib",
  "insightface",
  "onnxruntime",
  # todo: this should almost always be the headless version
  # the ecosystem really needs an alternative to cv2, the way they have packaged things is chaos
  "opencv-python",
  "timm",
]
optional-dependencies.torch25 = ["nunchaku==1.0.1+torch2.5; sys_platform != 'darwin'", "torch>=2.5,<2.6"]
optional-dependencies.torch26 = ["nunchaku==1.0.1+torch2.6; sys_platform != 'darwin'", "torch>=2.6,<2.7"]
optional-dependencies.torch27 = ["nunchaku==1.0.1+torch2.7; sys_platform != 'darwin'", "torch>=2.7,<2.8"]
optional-dependencies.torch28 = ["nunchaku==1.0.1+torch2.8; sys_platform != 'darwin'", "torch>=2.8,<2.9"]
# when working with pre-release torch, we don't know what the version string will be, so let's omit it
optional-dependencies.torchpre = ["nunchaku==1.0.1+torch2.10; sys_platform != 'darwin'"]
optional-dependencies.dev = [
  "pytest",
  "torchmetrics",
  "image_gen_aux@git+https://github.com/asomoza/image_gen_aux.git",
  "pyyaml",
  "pytest-rerunfailures",
  "pytest-asyncio==1.2.0",
  "comfyui@git+https://github.com/hiddenswitch/ComfyUI.git",
  # version < 3.13 required due to mediapipe
  "comfyui_controlnet_aux@git+https://github.com/AppMana/appmana-comfyui-nodes-controlnet-aux.git ;python_version < '3.13'",
]
urls.Repository = "https://github.com/nunchaku-tech/ComfyUI-nunchaku"
entry-points."comfyui.custom_nodes".nunchaku_comfyui = "nunchaku_comfyui"

[tool.setuptools]
package-dir = { "nunchaku_comfyui" = "." }
packages = [
  "nunchaku_comfyui",
  "nunchaku_comfyui.mixins",
  "nunchaku_comfyui.model_base",
  "nunchaku_comfyui.model_configs",
  "nunchaku_comfyui.models",
  "nunchaku_comfyui.nodes",
  "nunchaku_comfyui.nodes.lora",
  "nunchaku_comfyui.nodes.models",
  "nunchaku_comfyui.nodes.preprocessors",
  "nunchaku_comfyui.nodes.tools",
  "nunchaku_comfyui.wrappers",
]

[tool.setuptools.package-data]
"nunchaku_comfyui" = ["example_workflows/*.json"]
"nunchaku_comfyui.nodes.models" = ["nodes/models/configs/*.json"]

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120

per-file-ignores."tests/scripts/*.py" = ["E402"]
per-file-ignores."tests/legacy_scripts/*.py" = ["E402"]
per-file-ignores."__init__.py" = ["E402"]

[tool.isort]
profile = "black"
known_first_party = ["nunchaku"]
line_length = 120

[tool.comfy]
PublisherId = "lmxyy1999"
DisplayName = "ComfyUI-nunchaku"
Icon = "https://raw.githubusercontent.com/nunchaku-tech/nunchaku/96615bd93a1f0d2cf98039fddecfec43ce34cc96/assets/nunchaku.svg"

[tool.doc8]
max-line-length = 120
ignore-path = ["docs/_build"]
ignore = ["D000", "D001"]

[tool.rstcheck]
ignore_directives = ["tabs"]
ignore_messages = ["ERROR/3", "INFO/1"]

[tool.uv]
conflicts = [
  [
    { extra = "torch25" },
    { extra = "torch26" },
    { extra = "torch27" },
    { extra = "torch28" },
    { extra = "torchpre" },
    { extra = "compile" },
  ],
]

[[tool.uv.index]]
name = "nunchaku-wheels"
url = "https://raw.githubusercontent.com/AppMana/comfyui-nunchaku/refs/heads/main/pypi/nunchaku_index.html"
format = "flat"

[tool.uv.sources]
nunchaku = { index = "nunchaku-wheels" }

[[tool.uv.dependency-metadata]]
name = "nunchaku"
version = "1.0.1"
requires-dist = [
  "accelerate>=1.9",
  "diffusers>=0.35.1",
  "einops",
  "huggingface-hub>=0.34",
  "peft>=0.17",
  "protobuf",
  "sentencepiece",
  "torchvision>=0.20",
  # 4.53 shouldn't be used due to OpenTelemetry problems
  "transformers>=4.54,<4.57",
]
